[ADD] 출근, 퇴근, 정정

+++ working_time_log/slack/models.py
@@ -6,5 +6,5 @@ class Users(models.Model):
     username = models.CharField(max_length=40)
     entered_time = models.DateTimeField(null=True)
     exited_time = models.DateTimeField(null=True)
-    break_hours = models.IntegerField(default=0, null=True)
+    break_hours = models.DecimalField(default=0, max_digits=4, decimal_places=3, null=True)
     random_id = models.CharField(max_length=10)
